Edu/Notebook: Path planning UR5 robot Fixes #10
