--- edu/notebooks/rl/2D collision avoidance/jupyter/requirements_latest.txt
@@ -1,8 +1,0 @@
-mlpro[full]
-stable-baselines3
-mlpro-int-sb3
-gymnasium
-mlpro-int-gymnasium
-
-notebook
-jupyterlab
