--- edu/notebooks/rl/2D collision avoidance/jupyter/requirements_stable.txt
@@ -1,18 +1,0 @@
-dill==0.4.0
-multiprocess==0.70.18
-numpy==1.26.4
-matplotlib==3.9.0
-PySide6==6.9.0
-scipy==1.16.0
-pandas==2.1.3
-
-mlpro==2.1.0
-
-stable-baselines3==2.7.0
-mlpro-int-sb3==1.0.7
-
-gymnasium==1.2.0
-mlpro-int-gymnasium==1.0.4
-
-notebook==7.4.5
-jupyterlab==4.4.7
