--- edu/notebooks/rl/Path planning UR5 robot/jupyter/requirements_latest.txt
@@ -1,13 +1,0 @@
-mlpro[full]
-
-stable-baselines3
-mlpro-int-sb3
-
-gymnasium
-mlpro-int-gymnasium
-
-mujoco
-imageio
-
-notebook
-jupyterlab
