--- edu/notebooks/rl/Path planning UR5 robot/jupyter/requirements_stable.txt
@@ -1,26 +1,0 @@
-dill==0.4.0
-multiprocess==0.70.18
-numpy==1.26.4
-matplotlib==3.9.0
-PySide6==6.9.0
-scipy==1.16.0
-pandas==2.1.3
-
-mlpro==2.1.0
-
-stable-baselines3==2.3.2
-mlpro-int-sb3==1.0.7
-
-gymnasium==0.29.1
-mlpro-int-gymnasium==1.0.4
-
-mujoco==2.3.0
-imageio==2.34.0
-
-notebook==7.4.5
-jupyterlab==4.4.7
-
-
-# Warning if Python >= 3.12
-# (here a dummy package that intentionally does not exist)
-this-project-requires-python-3.10-or-3.11; python_version >= "3.12"
