--- edu/notebooks/rl/Path planning UR5 robot/jupyter/sim/__init__.py
@@ -1,6 +1,0 @@
-from gymnasium.envs.registration import register
-
-register(
-    id="UR5Env-v0",
-    entry_point="sim.ur5:UR5Env",
-)
