+++ edu/notebooks/rl/Path_planning_UR5_robot/jupyter/requirements_latest.off
@@ -1,0 +1,13 @@
+mlpro[full]
+
+stable-baselines3
+mlpro-int-sb3
+
+gymnasium
+mlpro-int-gymnasium
+
+mujoco
+imageio
+
+notebook
+jupyterlab
