+++ edu/notebooks/rl/Path_planning_UR5_robot/jupyter/requirements_stable.txt
@@ -1,0 +1,26 @@
+dill==0.4.0
+multiprocess==0.70.18
+numpy==1.26.4
+matplotlib==3.9.0
+PySide6==6.9.0
+scipy==1.16.0
+pandas==2.1.3
+
+mlpro==1.9.0
+
+stable-baselines3==2.3.2
+mlpro-int-sb3==1.0.2
+
+gymnasium==0.29.1
+mlpro-int-gymnasium==1.0.4
+
+mujoco==2.3.5
+imageio==2.34.0
+
+notebook==7.4.5
+jupyterlab==4.4.7
+
+
+# Warning if Python >= 3.12
+# (here a dummy package that intentionally does not exist)
+this-project-requires-python-3.10-or-3.11; python_version >= "3.12"
