+++ edu/notebooks/rl/Path_planning_UR5_robot/jupyter/sim/__init__.py
@@ -1,0 +1,6 @@
+from gymnasium.envs.registration import register
+
+register(
+    id="UR5Env-v0",
+    entry_point="sim.ur5:UR5Env",
+)
