+++ edu/notebooks/rl/Path_planning_UR5_robot/jupyter/sim/assets/env.xml
@@ -1,0 +1,27 @@
+<mujoco>
+    <worldbody>
+        <!--Target-->
+        <!--
+        <body name="safe_pick" pos="0 -.33 0.30" quat="0 1 0 0">
+            <geom conaffinity="0" contype="0" name="safe_pick" pos="0 0 0" rgba="0 0 0.8 1" size=".009" type="sphere"/>
+        </body>
+        -->
+        <body name="pick" pos="0 -.33 0.25" quat="0 1 0 0">
+            <geom conaffinity="0" contype="0" name="pick" pos="0 0 0" rgba="0.9 0.2 0.2 1" size=".009" type="sphere"/>
+        </body>
+        <!--
+        <body name="safe_place" pos=".3 -.33 0.30" quat="0 1 0 0">
+            <geom conaffinity="0" contype="0" name="safe_place" pos="0 0 0" rgba="0 0 0.8 1" size=".009" type="sphere"/>
+        </body>
+        -->
+        <body name="place" pos=".3 -.33 0.25" quat="0 1 0 0">
+            <geom conaffinity="0" contype="0" name="place" pos="0 0 0" rgba="0 0 0.8 1" size=".009" type="sphere"/>
+        </body>
+        <body name="target_allowance" pos="0 -0.33 0.25">
+            <geom conaffinity="0" contype="0" name="target_allowance" type="box" size="0.05 0.05 0.05" rgba="0.9 0.2 0.2 0.2" />
+        </body>
+        <body name="virtual_wall" pos="-0.05 -0.475 0.525">
+            <geom conaffinity="0" contype="0" name="virtual_wall" type="box" size="0.20 0.275 0.325" rgba="0.146 0.240 0.73 0.1" />
+        </body>
+    </worldbody>
+</mujoco>
